feat

+++ Exercises-Beecrowd/2540 - Impeachment do Lider/main.cpp
@@ -8,7 +8,7 @@ int main()
     int n ;
     while(cin>> n)
     {
-        int notaCorte = 0, favoraveis = 0, x = 0;
+        double notaCorte = 0, favoraveis = 0, x = 0;
         for(int i = 0 ; i < n ; i++)
         {
             cin>> x ;
@@ -29,10 +29,6 @@ int main()
         }
 
     }
-
-
-
-
 
     return 0;
 }
